防止规划所需的qp超过2147483648产生溢出，把相关的Int类型修改为Long类型

+++ app/src/main/java/com/ssttkkl/fgoplanningtool/data/item/Item.kt
@@ -8,19 +8,19 @@ import com.ssttkkl.fgoplanningtool.resources.ResourcesProvider
 
 @Entity(tableName = "Item")
 data class Item(@PrimaryKey val codename: String,
-                var count: Int) : Parcelable {
+                var count: Long) : Parcelable {
     val descriptor
         get() = ResourcesProvider.instance.itemDescriptors[codename]
 
     constructor(parcel: Parcel) : this(
             parcel.readString(),
-            parcel.readInt())
+            parcel.readLong())
 
     constructor(item: Item) : this(item.codename, item.count)
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeString(codename)
-        parcel.writeInt(count)
+        parcel.writeLong(count)
     }
 
     override fun describeContents(): Int {
